+++ src/lib.rs
@@ -233,6 +233,7 @@ pub mod bitreader;
 pub mod cabac;
 pub mod ctx_init;
 pub mod hrd;
+pub mod inter_pred;
 pub mod intra_pred;
 pub mod nal;
 pub mod pps;